requirements.txt
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
jsonschema==4.22.0
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
pathable==0.4.3
PyYAML==6.0.1
referencing==0.31.1
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
urllib3==2.2.2
